+++ src/components/composite/agGridTable/CustomAgGridTable.tsx
@@ -10,8 +10,8 @@ import { useFieldArray, useFormContext } from 'react-hook-form';
 import { Box, useTheme } from '@mui/material';
 import type { CellEditingStoppedEvent, ColumnState, SortChangedEvent } from 'ag-grid-community';
 import { BottomRightButtons } from './BottomRightButtons';
-import { FieldConstants } from '../../../../utils/constants/fieldConstants';
-import { CustomAGGrid, type CustomAGGridProps } from '../../customAGGrid';
+import { FieldConstants } from '../../../utils/constants/fieldConstants';
+import { CustomAGGrid, type CustomAGGridProps } from '../customAGGrid';
 
 const style = (customProps: any) => ({
     grid: (theme: any) => ({
