+++ src/components/composite/customAGGrid/cell-renderers.tsx
@@ -10,7 +10,7 @@ import { ReactNode, useRef, useState } from 'react';
 import { ICellRendererParams } from 'ag-grid-community';
 import { CustomCellRendererProps } from 'ag-grid-react';
 import { useIntl } from 'react-intl';
-import { CustomTooltip } from '../tooltip/CustomTooltip';
+import { CustomTooltip } from '../../ui/tooltip/CustomTooltip';
 import { isBlankOrEmpty, mergeSx, MuiStyles } from '../../../utils';
 
 const styles = {
