+++ src/components/composite/filter/explicitNaming/ExplicitNamingFilterForm.tsx
@@ -14,10 +14,10 @@ import { v4 as uuid4 } from 'uuid';
 import type { UUID } from 'node:crypto';
 import { FieldConstants } from '../../../../utils/constants/fieldConstants';
 import yup from '../../../../utils/yupConfig';
-import { CustomAgGridTable } from '../../../ui/reactHookForm/agGridTable/CustomAgGridTable';
+import { CustomAgGridTable } from '../../agGridTable/CustomAgGridTable';
 import { SelectInput } from '../../../ui/reactHookForm/selectInputs/SelectInput';
 import { isInjection } from '../../../../utils/types/equipmentTypes';
-import { NumericEditor } from '../../../ui/reactHookForm/agGridTable/cellEditors/numericEditor';
+import { NumericEditor } from '../../agGridTable/cellEditors/numericEditor';
 import { InputWithPopupConfirmation } from '../../../ui/reactHookForm/selectInputs/InputWithPopupConfirmation';
 import { toFloatOrNullValue } from '../../../ui/reactHookForm/utils/functions';
 import { DISTRIBUTION_KEY } from '../constants/FilterConstants';
