+++ src/components/composite/index.ts
@@ -11,3 +11,5 @@ export * from './filter';
 export * from './grid';
 export * from './muiTable';
 export * from './reactQueryBuilder';
+export * from './agGridTable';
+export * from './customAGGrid';
