+++ src/components/ui/index.ts
@@ -19,4 +19,3 @@ export * from './resizablePanels';
 export * from './csvDownloader';
 export * from './treeViewFinder';
 export * from './directoryItemSelector';
-export * from './customAGGrid';
